Update to gamescope 3.9.2
This also adds stb as a required in-repo depenency for gamescope. It's
not cloned automatically by flatpak-builder because it's a Meson
git-wrap file, not a git submodule.

Instead, we clone it ourselves to the expected subprojects directory.

However, when using the "meson submodule download" command locally to
download its source or when building it in a non-sandboxed environment
with Internet access, git-wrap also makes it appear as a Meson
project. This is because gamescope provides a meson.build file for it
in the following directory: subprojects/packagefiles/stb/meson.build

Since gamescope's buildsystem actually depends on it appearing like a
Meson project, we must make sure it remains such, so we manually copy
the content of the subprojects/packagefiles/stb directory to its
subproject source directory.

In addition, because Meson wraps files are present, we also set
--wrap-mode=nodownload to prevent Meson from attempting to download
sources on its own.